2010-06-22 16:50 UTC+0200 Viktor Szakats (harbour.01 syenar.hu)   * src/vm/dynsym.c   * include/hbapi.h     * Minimal cleanup: using HB_LONG in place of long, just to make       this type usage easier to find in the future.
  * src/vm/hvm.c
    + Added cast to pacify long time msvc warning

  * src/vm/fm.c
    + Added pragmas to silence msvc64 warnings in dlmalloc (foreign) code.

  * src/vm/extend.c
    * Formatting.

+++ harbour/src/vm/hvm.c
@@ -10696,7 +10696,7 @@ HB_BOOL hb_xvmModulusByInt( HB_LONG lDivisor )
          }
       }
       else if( HB_IS_NUMINT( pValue ) )
-         hb_itemPutND( pValue, HB_ITEM_GET_NUMINTRAW( pValue ) % lDivisor );
+         hb_itemPutND( pValue, ( double ) ( HB_ITEM_GET_NUMINTRAW( pValue ) % lDivisor ) );
       else
          hb_itemPutND( pValue, fmod( hb_itemGetND( pValue ), lDivisor ) );
